requirements.txt
torch
torchvision
transformers
open_clip_torch
spacy
opencv-python
Pillow
streamlit